Added Calculation service, Tactic->ExtractPromotion->ExtractProductFromPromotion->Return Uplift+Price+Revenue

+++ tpm-backend/tpm-backend/src/main/java/com/kirti/tpm/entity/Tactic.java
@@ -31,7 +31,7 @@ public class Tactic {
     @Enumerated(EnumType.STRING)
     private TacticType tacticType;
 
-    private int discount;
+    private double discount;
     private int perProduct;
     private double lumpsum;
